ext/VisualizationExt.jl: Make CairoMakie optional and plot_diagnostics as extension.
