+++ src/LineageCollapse.jl
@@ -4,12 +4,11 @@ module LineageCollapse
     using ProgressMeter
     using Clustering
     using BioSequences
-    using CairoMakie
 
     export load_data, preprocess_data, process_lineages, plot_diagnostics
+    function plot_diagnostics(df::DataFrame) end
 
     include("data_loading.jl")
     include("preprocessing.jl")
     include("lineage_processing.jl")
-    include("visualization.jl")
 end
